made creating the driver work on Windows
this commit adds a variable, `ob-ipython-python-executable-search-string` that is set to "python.exe" if the system type is "windows-nt". Otherwise, it is set to "python". This change fixes https://github.com/gregsexton/ob-ipython/issues/4#issuecomment-111040138

However, I have only been able to test on windows, so someone should test on Linux/OSX to be sure it doesn't mess anything up there.