support both IPython3 and Jupyter by conditional import
the code first tries to import the Jupyter libraries, failing that
the IPython 3 libraries are imported. All that really changed are
module names.

This simple compatibility workaround may not work
in the long term if Jupyter code diverges significantly from the current state. But by that time IPython 3 will be hopefully obsolete.