implement remote kernel connection through SSH
This uses the --ssh option of the ipython console app to create the necessary
tunnels and create a new connection file with "-ssh" added (because it may use
different ports). Therefore, the REPL is created before the driver. The
connection file naming convention of "emacs-*.ssh" is abandoned in order to
support arbitrary remote kernel connection files. Fixes #11.