Enable session-dedicated REPLs.
Change ob-ipython--create-repl so that when a named session is used
(i.e. not "default") a dedicated repl is created for that
session. This enables state in a named session to be accessbile within
the dedicated REPL. When the default session is used, the previous
behaviour of creating a global Python process is preserved.